Update to Linux Instructions (#2055)
* Update to Linux Instructions

Put db creation in a code block to prevent copy and paste causing wrong single quote to be copied.
Fixed typo  " .*" to "*.*"
Added " WITH GRANT OPTION" per MariaDB documentation

* Update mysql.md

I did not mean to omit 'rocketmapdb' just replace the missing wild card.

* Update mysql.md

OK, it was not a missing wild card (although that will make it work), it was the spaces that was the problem... "rocketmapdb . * instead of "rocketmapdb.*"

(cherry picked from commit fd2cad1)